+++ tests/mosaic/gpu_dialect_test.py
@@ -2849,6 +2849,30 @@ class DialectLoweringTest(MosaicGpuTest):
       self.assertTrue(any(f.endswith(".before_layout_inference.txt") for f in files))
       self.assertTrue(any(f.endswith(".after_layout_inference.txt") for f in files))
 
+  @jtu.thread_unsafe_test()  # Modifies ``os.environ``.
+  def test_dump_constraint_system(self):
+    # TODO(bchetioui): Remove this once minimum jaxlib version is 0.11.2.
+    if not hasattr(mgpu.dialect.DumpOptions(), "constraint_system"):
+      self.skipTest("Test requires jaxlib >= 0.11.2")
+    def body(_, src, dst, scratch):
+      del scratch
+      mgpu.dialect.vector_store(mgpu.dialect.vector_load(src), dst)
+    shape = (128, 128)
+
+    with tempfile.TemporaryDirectory() as dump_dir:
+      with jtu.set_env(MOSAIC_GPU_DUMP_TO=dump_dir):
+        mgpu.as_gpu_kernel(
+            body,
+            grid=(1, 1, 1),
+            block=(128, 1, 1),
+            in_shape=jax.ShapeDtypeStruct(shape, jnp.float32),
+            out_shape=jax.ShapeDtypeStruct(shape, jnp.float32),
+            smem_scratch_shape=[],
+            thread_semantics=mgpu.LoweringSemantics.Warpgroup,
+        )
+      files = os.listdir(dump_dir)
+      self.assertTrue(any(f.endswith(".constraint_system.txt") for f in files))
+
 
 if hp is not None:
   @hps.composite